requirements.txt
asgiref==3.8.1
Django==5.2.1
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Markdown==3.8
PyJWT==2.9.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
